use sprintf instead of multiple lines

+++ arduino/main/emg.h
@@ -4,6 +4,6 @@
 #include <ArduinoBLE.h>
 
 void setup_emg();
-double read_emg(int port);
+float read_emg(int port);
 
 #endif
